requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
aiohttp==3.9.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
prefect==2.14.11
robotframework==6.1.1
robotframework-requests==0.9.4
griffe==0.36.6
python-dotenv==1.1.0